refactor(backend): split aggregation sql_fragments by domain
Decomposes the 748-LOC aggregations/sql_fragments.py into a
sub-package at the same import path:

  aggregations/sql_fragments/
    __init__.py         (96)   re-exports every public name
    paths.py            (24)   JSONB path constants (VD_ID, VD_EXTENSIONS, ...)
    classification.py   (315)  VARIANT_TYPE_CLASSIFICATION_SQL + CASEs
    ctes.py             (168)  PHENOPACKET_VARIANT_LINK_CTE, UNIQUE_VARIANTS_CTE, etc.
    protein_domain.py   (148)  HNF1B domain classification helpers

Every file is under 500 LOC. Public import path unchanged:

    from app.phenopackets.routers.aggregations.sql_fragments import (
        UNIQUE_VARIANTS_CTE,
        get_pending_variants_query,
        HNF1B_PROTEIN_DOMAINS,
        # ...
    )

All 20+ existing import sites (admin sync_service, sync_variants_routes,
survival handlers, test_survival_protein_domain) continue to work via
the package-level re-export in __init__.py.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
